backend/requirements.txt
Flask==2.2.3
Flask-Cors==3.0.10
Flask-JWT-Extended==4.4.4
Flask-SQLAlchemy==3.0.3
SQLAlchemy==2.0.4
Werkzeug==2.2.3
bcrypt==4.0.1
cryptography==39.0.1
gunicorn==20.1.0
python-dotenv==1.0.0
requests==2.28.2
google-cloud-storage==2.7.0
google-auth==2.16.2
Pillow==9.4.0
pytest==7.2.2
pytest-flask==1.2.0
selenium==4.8.2
pycryptodome==3.17.0
pyotp==2.8.0
qrcode==7.4.2
email-validator==1.3.1
marshmallow==3.19.0


flask-socketio==5.3.4
python-engineio==4.4.1
python-socketio==5.8.0
eventlet==0.33.3